Consolidated ErrorResponse and ReplyHelper
